refactor: move privacy policy to pages directory
